+++ CHANGES.md
@@ -6,6 +6,9 @@
 
 * A process is not considered disabled anymore when process rules don't allow any candidate **Supvisors** instance.
 
+* When `psutil` is not installed on a host, the statistics-related options of the Process and Host pages
+  of the Web UI are not displayed, just as if the option `stats_enabled` was set to `False`.
+
 
 ## 0.14 (2022-05-01)